api/src/requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.6.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
python-decouple==3.3
pytz==2020.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.4.0
